refactor: move APIs to separate directory
