+++ openassessment/xblock/apis/ora_data_accessor.py
@@ -1,11 +1,11 @@
 """API Data wrapper for exposed APIs within ORA XBlock"""
-from openassessment.xblock.ora_config_api import ORAConfigAPI
-from openassessment.xblock.submissions.submissions_api import SubmissionAPI
-from openassessment.xblock.workflow_api import WorkflowAPI
-from openassessment.xblock.assessments.peer_assessment_api import PeerAssessmentAPI
-from openassessment.xblock.assessments.self_assessment_api import SelfAssessmentAPI
-from openassessment.xblock.assessments.staff_assessment_api import StaffAssessmentAPI
-from openassessment.xblock.assessments.student_training_api import StudentTrainingAPI
+from openassessment.xblock.apis.ora_config_api import ORAConfigAPI
+from openassessment.xblock.apis.submissions.submissions_api import SubmissionAPI
+from openassessment.xblock.apis.workflow_api import WorkflowAPI
+from openassessment.xblock.apis.assessments.peer_assessment_api import PeerAssessmentAPI
+from openassessment.xblock.apis.assessments.self_assessment_api import SelfAssessmentAPI
+from openassessment.xblock.apis.assessments.staff_assessment_api import StaffAssessmentAPI
+from openassessment.xblock.apis.assessments.student_training_api import StudentTrainingAPI
 
 class ORADataAccessor:
     def __init__(self, block):
